[support cctv source & clean code]
